level: tidy up level apis
