+++ src/meson.build
@@ -312,12 +312,24 @@ common_sources = [
   'trx/game/lara/state/swim.c',
   'trx/game/lara/vehicle.c',
   'trx/game/level/common.c',
-  'trx/game/level/loader.c',
-  'trx/game/level/reader.c',
-  'trx/game/level/reader_tr1.c',
-  'trx/game/level/reader_tr2.c',
-  'trx/game/level/reader_tr3.c',
-  'trx/game/level/rooms.c',
+  'trx/game/level/context.c',
+  'trx/game/level/finalize/animations.c',
+  'trx/game/level/finalize/gameplay_objects.c',
+  'trx/game/level/finalize/render_assets.c',
+  'trx/game/level/finalize/rooms.c',
+  'trx/game/level/format/format_tr1.c',
+  'trx/game/level/format/format_tr2.c',
+  'trx/game/level/format/format_tr3.c',
+  'trx/game/level/format/pipeline.c',
+  'trx/game/level/pipeline.c',
+  'trx/game/level/sections/anims.c',
+  'trx/game/level/sections/audio.c',
+  'trx/game/level/sections/cinematics.c',
+  'trx/game/level/sections/meshes.c',
+  'trx/game/level/sections/objects.c',
+  'trx/game/level/sections/pathing.c',
+  'trx/game/level/sections/rooms.c',
+  'trx/game/level/sections/textures.c',
   'trx/game/level/settings.c',
   'trx/game/los.c',
   'trx/game/lua/assault_stats.c',
