+++ src/trx/game/inject/common.h
@@ -14,7 +14,7 @@ INJECTION_MESH_META Inject_GetRoomMeshMeta(int32_t room_index);
 int32_t Inject_GetDataCount(INJECTION_DATA_TYPE type);
 int32_t Inject_GetMaxStaticObject3DId(void);
 int32_t Inject_GetMaxStaticObject2DId(void);
-LEVEL_INFO Inject_GetCachedInfo(void);
+LEVEL_CONTEXT_INFO Inject_GetCachedInfo(void);
 
 void Inject_RegisterPaletteMap(const uint16_t *palette_map, int32_t size);
 uint16_t Inject_GetPaletteIndex(uint16_t index);
@@ -27,8 +27,8 @@ void Inject_RegisterHandler(
     void (*handle_func)(const INJECTION_CONTEXT *, INJECTION_CHUNK chunk));
 
 #define REGISTER_INJECT_TESTER(test_type, test_func)                           \
-    __attribute__((                                                            \
-        constructor)) static void M_RegisterInjectTester##test_type(void)      \
+    __attribute__((constructor)) static void                                   \
+    M_RegisterInjectTester##test_type(void)                                    \
     {                                                                          \
         Inject_RegisterTester(test_type, test_func);                           \
     }
