+++ src/trx/game/inject/data/meshes.c
@@ -1,5 +1,7 @@
 #include <trx/debug.h>
 #include <trx/game/inject.h>
+#include <trx/game/level/context.h>
+#include <trx/game/level/sections/append.h>
 #include <trx/memory.h>
 
 static void M_HandleMeshData(
@@ -31,9 +33,9 @@ static void M_HandleMeshData(
 
         case IDT_OBJECT_MESHES: {
             ASSERT(mesh_indices != nullptr);
-            Level_AppendObjectMeshes(
+            Level_Section_AppendObjectMeshes(
                 mesh_ptr_count, mesh_indices, chunk.injection->fp);
-            LEVEL_INFO *const info = Level_GetInfo();
+            LEVEL_CONTEXT_INFO *const info = Level_Context_GetInfo();
             info->mesh_ptr_count += mesh_ptr_count;
             break;
         }
