+++ src/trx/game/level.h
@@ -1,7 +1,7 @@
 #pragma once
 
 #include <trx/game/level/common.h>
-#include <trx/game/level/loader.h>
-#include <trx/game/level/reader.h>
+#include <trx/game/level/context.h>
+#include <trx/game/level/format/format.h>
+#include <trx/game/level/pipeline.h>
 #include <trx/game/level/settings.h>
-#include <trx/game/level/types.h>
