+++ src/trx/game/level/pipeline.h
@@ -1,0 +1,9 @@
+#pragma once
+
+#include <trx/game/game_flow/types.h>
+#include <trx/game/level/context.h>
+#include <trx/game/level/finalize.h>
+#include <trx/game/level/sections/append.h>
+#include <trx/game/level/sections/read.h>
+
+void Level_Pipeline_Load(const GF_LEVEL *level);
