--- src/trx/game/level/reader_tr1.h
@@ -1,7 +1,0 @@
-#pragma once
-
-#include <trx/game/level/reader.h>
-
-extern const LEVEL_LOADER g_LevelLoaderTR1;
-extern const LEVEL_LOADER g_LevelLoaderTR1X;
-extern const LEVEL_LOADER g_LevelLoaderTR1DemoPC;
