--- src/trx/game/level/reader_tr2.h
@@ -1,6 +1,0 @@
-#pragma once
-
-#include <trx/game/level/reader.h>
-
-extern const LEVEL_LOADER g_LevelLoaderTR2;
-extern const LEVEL_LOADER g_LevelLoaderTR2X;
