--- src/trx/game/level/reader_tr3.h
@@ -1,5 +1,0 @@
-#pragma once
-
-#include <trx/game/level/reader.h>
-
-extern const LEVEL_LOADER g_LevelLoaderTR3;
